+++ packages/editor/src/components/parts/common/path/PathCollapsible.tsx
@@ -1,15 +1,15 @@
 import { ReactNode } from 'react';
 import { PathContext } from '../../../../context';
-import CollapsiblePart, { CollapsiblePartProps } from '../../../widgets/collapsible/CollapsiblePart';
+import Collapsible, { CollapsibleProps } from '../../../widgets/collapsible/Collapsible';
 import { SchemaKeys } from '@axonivy/inscription-protocol';
 
-type PathCollapsibleProps = CollapsiblePartProps & {
+type PathCollapsibleProps = CollapsibleProps & {
   path: SchemaKeys;
   children: ReactNode;
 };
 
 export const PathCollapsible = ({ path: location, children, ...props }: PathCollapsibleProps) => (
   <PathContext path={location}>
-    <CollapsiblePart {...props}>{children}</CollapsiblePart>
+    <Collapsible {...props}>{children}</Collapsible>
   </PathContext>
 );
