+++ packages/editor/src/components/parts/task/expiry/ExpiryPart.tsx
@@ -12,7 +12,7 @@ const ExpiryPart = () => {
   const timeoutFieldset = useFieldset();
 
   return (
-    <PathCollapsible collapsibleLabel='Expiry' defaultOpen={isTimeout} path='expriy'>
+    <PathCollapsible label='Expiry' defaultOpen={isTimeout} path='expriy'>
       <PathFieldset label='Timeout' {...timeoutFieldset.labelProps} path='timeout'>
         <ScriptInput
           value={expiry.timeout}
