+++ packages/editor/src/components/parts/task/options/PersistPart.tsx
@@ -1,13 +1,13 @@
-import { Checkbox, CollapsiblePart } from '../../../../components/widgets';
+import { Checkbox, Collapsible } from '../../../../components/widgets';
 import { useTaskPersistData } from './useTaskOptionsData';
 
 const PersistPart = () => {
   const { config, updatePersist } = useTaskPersistData();
 
   return (
-    <CollapsiblePart collapsibleLabel='Options' defaultOpen={config.persist}>
+    <Collapsible label='Options' defaultOpen={config.persist}>
       <Checkbox label='Persist task on creation' value={config.persist} onChange={updatePersist} />
-    </CollapsiblePart>
+    </Collapsible>
   );
 };
