+++ packages/editor/src/components/widgets/collapsible/Collapsible.css
